requirements.txt
arabic-reshaper==3.0.0
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
Pillow==10.1.0
pyTelegramBotAPI==4.14.1
python-decouple==3.8
requests==2.31.0
soupsieve==2.5
urllib3==2.1.0
